Fix marshaling of floats for credit hours, also refactor some structs out of terms and into their relevant services.

+++ courses.go
@@ -7,18 +7,18 @@ import (
 )
 
 type Course struct {
-	Id                       int    `json:"id"`
-	Term                     string `json:"term"`
-	CourseRegistrationNumber string `json:"courseReferenceNumber"`
-	Subject                  string `json:"subject"`
-	SectionNumber            string `json:"sequenceNumber"`
-	Title                    string `json:"courseTitle"`
-	IsOpen                   bool   `json:"openSection"`
-	CrossList                string `json:"crossList"`
-	CrossListCapacity        int    `json:"crossListCapacity"`
-	CrossListAvailable       int    `json:"crossListAvailable"`
-	CreditHourHigh           int    `json:"creditHourHigh"`
-	CreditHourLow            int    `json:"creditHourLow"`
+	Id                       int     `json:"id"`
+	Term                     string  `json:"term"`
+	CourseRegistrationNumber string  `json:"courseReferenceNumber"`
+	Subject                  string  `json:"subject"`
+	SectionNumber            string  `json:"sequenceNumber"`
+	Title                    string  `json:"courseTitle"`
+	IsOpen                   bool    `json:"openSection"`
+	CrossList                string  `json:"crossList"`
+	CrossListCapacity        int     `json:"crossListCapacity"`
+	CrossListAvailable       int     `json:"crossListAvailable"`
+	CreditHourHigh           float64 `json:"creditHourHigh"`
+	CreditHourLow            float64 `json:"creditHourLow"`
 }
 
 func FetchTermDepartmentCourses(t *Term, d *Department) (string, error) {
